Chat_With_CSV_Via_Langchain/requirements.txt
openai==1.6.1
langchain==0.0.354
langchain-experimental==0.0.47
streamlit
python-dotenv
tabulate
# dotenv